+++ src/Exceptionist.php
@@ -21,7 +21,6 @@ use Tools\Exception\FileNotExistsException;
 use Tools\Exception\KeyNotExistsException;
 use Tools\Exception\NotReadableException;
 use Tools\Exception\PropertyNotExistsException;
-use function Symfony\Component\String\u;
 
 /**
  * Exceptionist
@@ -60,7 +59,7 @@ class Exceptionist
     public static function __callStatic($name, array $arguments)
     {
         //Gets the PHP function name
-        $name = (string)u($name)->snake();
+        $name = uncamelcase($name);
         if (!function_exists($name)) {
             trigger_error(sprintf('Function `%s()` does not exist', $name));
         }